Apple Sales Reporter
Added a sleep after clicking the download button in order to ensure that we start downloading the file. Since the check on the file system for *.part was was occurring before firefox even started to write the file to disk.

Updated the bot to take a more stable path to the report file.

Signed-off-by: Justin Jeffress <sleepydeveloper@gmail.com>