booktrope-modules.rb initial commit
ATTENTION: This file was missed in a commit since it wasn't properly added to the .gitignore. This necessary file should have been checked into SHA:

a437c73e7f81a7636b593b995915567beb117347

code from the above commit up to this one should not be used in production since it will lack this file and may fail to run.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>