TTR-97 - Scrape Secret Cravings for Books with 30 or more reviews
Built a lightweight framework for scraping Amazon using the Advanced Search page.  This uses a
regular browser (not headless) and outputs to a CSV file since it's only intended for one-time
uses.  The framework should be very easy to use to perform similar searches and scrapes though
there aren't a ton of fields returned.