TTR-215 - Fix the collecting of reviews and star rating
Added some headers to the downloadData function which seems to improve
our chances of pulling data from Amazon without an error.

Changed some formatting that my IDE had issue with due to a lack of
whitespace.

Noticed an issue with the batching where it only iterated through the
first batch of ten items and then done is set to true when it shouldn't
be.

https://teamtrope.atlassian.net/browse/TTR-215